Refactor data types from float to double in plotting functions and structures for improved precision
- Changed vector types from std::vector<float> to std::vector<double> in examples and plotgen files.
- Updated function parameters and return types to use double instead of float for consistency.
- Enhanced axis limit calculations and histogram data handling to accommodate double precision.
- Adjusted symbol drawing functions to utilize double for size and radius calculations.

+++ include/plotgen.h
@@ -18,25 +18,25 @@ class PlotGen {
 public:
     struct Style {
         sf::Color color;
-        float thickness;
+        double thickness;
         std::string line_style;
         std::string legend;
         std::string symbol_type; // "none", "circle", "square", "triangle", "diamond", "star"
-        float symbol_size;      // Symbol size in pixels
+        double symbol_size;      // Symbol size in pixels
 
         Style(
             sf::Color color_ = sf::Color::Blue,
-            float thickness_ = 2.0f,
+            double thickness_ = 2.0f,
             const std::string& line_style_ = "solid",
             const std::string& legend_ = "",
             const std::string& symbol_type_ = "none",
-            float symbol_size_ = 6.0f
+            double symbol_size_ = 6.0f
         );
     };
 
     struct Figure {
         std::string title, xlabel, ylabel;
-        float xmin = -10, xmax = 10, ymin = -10, ymax = 10;
+        double xmin = -10, xmax = 10, ymin = -10, ymax = 10;
         bool show_leg = true;
         bool show_major_grid = false; // Option to display major grid
         bool show_minor_grid = false; // Option to display minor grid
@@ -45,9 +45,9 @@ public:
         bool is_polar = false; // Indicates if the graph is in polar coordinates
         bool equal_axes = false; // Option for axes of the same dimension
         struct Curve {
-            std::vector<float> x, y;
+            std::vector<double> x, y;
             Style style;
-            float bar_width_ratio = 0.9f; // Field to store width ratio
+            double bar_width_ratio = 0.9f; // Field to store width ratio
         };
         std::vector<Curve> curves;
         std::vector<std::string> curve_types;
@@ -63,21 +63,21 @@ public:
     void set_title(Figure& fig, const std::string& title);
     void set_xlabel(Figure& fig, const std::string& label);
     void set_ylabel(Figure& fig, const std::string& label);
-    void set_axis_limits(Figure& fig, float xmin, float xmax, float ymin, float ymax);
-    void set_polar_axis_limits(Figure& fig, float max_radius);
+    void set_axis_limits(Figure& fig, double xmin, double xmax, double ymin, double ymax);
+    void set_polar_axis_limits(Figure& fig, double max_radius);
     void show_legend(Figure& fig, bool show);
     void grid(Figure& fig, bool major = true, bool minor = false);
     void set_grid_color(Figure& fig, sf::Color major_color, sf::Color minor_color);
     void set_equal_axes(Figure& fig, bool equal = true);
 
     // 2D curve plotting
-    void plot(Figure& fig, const std::vector<float>& x, const std::vector<float>& y, const Style& style = Style());
+    void plot(Figure& fig, const std::vector<double>& x, const std::vector<double>& y, const Style& style = Style());
 
     // Histogram
-    void hist(Figure& fig, const std::vector<float>& data, int bins = 10, const Style& style = Style(), float bar_width_ratio = 0.9f);
+    void hist(Figure& fig, const std::vector<double>& data, int bins = 10, const Style& style = Style(), double bar_width_ratio = 0.9f);
 
     // Polar plot
-    void polar_plot(Figure& fig, const std::vector<float>& theta, const std::vector<float>& r, const Style& style = Style());
+    void polar_plot(Figure& fig, const std::vector<double>& theta, const std::vector<double>& r, const Style& style = Style());
 
     // Display and render
     void show();
@@ -99,14 +99,14 @@ private:
 
     // Rendering methods
     void render();
-    void draw_axes(const Figure& fig, float w, float h);
-    void draw_grid(const Figure& fig, float w, float h);
-    void draw_polar_grid(const Figure& fig, float w, float h);
-    void draw_curve(const Figure& fig, const Figure::Curve& curve, float w, float h);
-    void draw_histogram(const Figure& fig, const Figure::Curve& curve, float w, float h);
-    sf::Color getColorFromHeight(float height);
-    sf::Vector2f to_screen(const Figure& fig, float x, float y, float w, float h) const;
-    void draw_text(const Figure& fig, float w, float h);
-    void draw_symbol(const sf::Vector2f& position, const std::string& symbol_type, float size, const sf::Color& color);
+    void draw_axes(const Figure& fig, double w, double h);
+    void draw_grid(const Figure& fig, double w, double h);
+    void draw_polar_grid(const Figure& fig, double w, double h);
+    void draw_curve(const Figure& fig, const Figure::Curve& curve, double w, double h);
+    void draw_histogram(const Figure& fig, const Figure::Curve& curve, double w, double h);
+    sf::Color getColorFromHeight(double height);
+    sf::Vector2f to_screen(const Figure& fig, double x, double y, double w, double h) const;
+    void draw_text(const Figure& fig, double w, double h);
+    void draw_symbol(const sf::Vector2f& position, const std::string& symbol_type, double size, const sf::Color& color);
 };
